src, include: robotics: moved common part to shared_type
